Walk Movement

+++ SmashUE_ClearedVersion-master/Source/SmashUE/Public/Characters/States/SmashCharacterStateWalk.h
@@ -26,5 +26,5 @@ protected:
 	UAnimMontage* WalkMontage;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite)
-	int MoveSpeedMax;
+	float MoveSpeedMax;
 };
